WIP: save data with --xfail / -x
When using `-x` to abort the test run on the first failure,
pytest-testmon would previously not store the collected data: the
`pytest_keyboard_interrupt` hook that is used for this gets also called
for pytest's internal `Interrupted` exception, which is a subclass of
`KeyboardInterrupt` (https://github.com/pytest-dev/pytest/issues/1865).

This patch changes it to use the result in the `pytest_runtest_protocol`
method to check for `KeyboardInterrupt` there explicitly.

This makes the `pytest_keyboard_interrupt` obsolete, but I've left it
with some asserts for now.